requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt
msal
pyodbc
python-dotenv